Działający projekt
Na razie bez merge'a logowania i rejestracji
