+++ .metadata
@@ -4,7 +4,7 @@
 # This file should be version controlled and should not be manually edited.
 
 version:
-  revision: 60bd88df915880d23877bfc1602e8ddcf4c4dd2a
+  revision: 27321ebbad34b0a3fafe99fac037102196d655ff
   channel: stable
 
 project_type: app
