--- lib/core/models/cart_item.dart
@@ -1,17 +1,0 @@
-import 'package:flutter/foundation.dart';
-
-class CartItem {
-  final String id;
-  final String title;
-  final int quantity;
-  final double price;
-  final String imageUrl;
-
-  CartItem({
-    @required this.id,
-    @required this.title,
-    @required this.quantity,
-    @required this.price,
-    @required this.imageUrl,
-  });
-}
