--- lib/core/models/order_item.dart
@@ -1,16 +1,0 @@
-import 'package:firebase_shop_app/core/models/cart_item.dart';
-import 'package:flutter/foundation.dart';
-
-class OrderItem {
-  final String id;
-  final double amount;
-  final List<CartItem> products;
-  final DateTime dateTime;
-
-  OrderItem({
-    @required this.id,
-    @required this.amount,
-    @required this.products,
-    @required this.dateTime,
-  });
-}
