--- lib/core/providers/product_model_provider.dart
@@ -1,24 +1,0 @@
-import 'package:flutter/foundation.dart';
-
-class ProductModelProvider with ChangeNotifier {
-  final String id;
-  final String title;
-  final String description;
-  final double price;
-  final String imageUrl;
-  bool isFavorite;
-
-  ProductModelProvider({
-    @required this.id,
-    @required this.title,
-    @required this.description,
-    @required this.price,
-    @required this.imageUrl,
-    this.isFavorite = false,
-  });
-
-  void changeFavoriteStatus() {
-    isFavorite = !isFavorite;
-    notifyListeners();
-  }
-}
