--- lib/utils/core/shared_prefrences.dart
@@ -1,11 +1,0 @@
-import 'package:shared_preferences/shared_preferences.dart';
-
-addStringToSF(String key, String value) async {
-  SharedPreferences prefs = await SharedPreferences.getInstance();
-  prefs.setString(key, value);
-}
-
-addIntToSF(String key, int value) async {
-  SharedPreferences prefs = await SharedPreferences.getInstance();
-  prefs.setInt(key, value);
-}
