+++ lib/utils/view/theme_manager.dart
@@ -1,5 +1,5 @@
-import 'package:flutter/material.dart';
 import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
 
 class ThemeManager {
   static String appName = "";
@@ -9,26 +9,21 @@ class ThemeManager {
   static Color darkPrimary = Colors.black;
   static Color primaryColor = Color(0xFFFA6442);
   static Color lightAccent = Color(0xFFFA6442);
-  static Color darkAccent =  Color(0xFFFA6442);
+  static Color darkAccent = Color(0xFFFA6442);
+
 //  static Color lightBG = Color(0xfffcfcff);
   static Color lightBG = Colors.grey.shade50;
   static Color darkBG = Colors.black;
   static Color badgeColor = Color(0xFFFA6442);
 
-
-
-
   static ThemeData lightTheme = ThemeData(
-
-
     fontFamily: 'cupertino',
     backgroundColor: lightBG,
     primaryColor: lightPrimary,
     scaffoldBackgroundColor: lightPrimary,
-      appBarTheme: AppBarTheme(
-        color: const Color(0xFFFA6442),
-
-      ),
+    appBarTheme: AppBarTheme(
+      color: const Color(0xFFFA6442),
+    ),
     textTheme: TextTheme(
       headline1: TextStyle(
           color: Colors.black, fontSize: 22.0, fontWeight: FontWeight.w800),
@@ -45,8 +40,8 @@ class ThemeManager {
       buttonColor: Color(0xFFFA6442),
       height: 50.0,
       textTheme: ButtonTextTheme.primary,
-    ), colorScheme: ColorScheme.fromSwatch().copyWith(secondary: lightAccent), textSelectionTheme: TextSelectionThemeData(cursorColor: lightAccent),
-
-
+    ),
+    colorScheme: ColorScheme.fromSwatch().copyWith(secondary: lightAccent),
+    textSelectionTheme: TextSelectionThemeData(cursorColor: lightAccent),
   );
 }
