--- lib/utils/view/ui_spacing_helper.dart
@@ -1,34 +1,0 @@
-import 'package:flutter/material.dart';
-
-/// Contains useful consts to reduce boilerplate and duplicate code
-class UISpacingHelper {
-  // Vertical spacing constants. Adjust to your liking.
-  static const double _VerticalSpaceXSmall = 4.0;
-  static const double _VerticalSpaceSmall = 10.0;
-  static const double _VerticalSpaceMedium = 20.0;
-  static const double _VerticalSpaceLarge = 60.0;
-
-  // Vertical spacing constants. Adjust to your liking.
-  static const double _HorizontalSpaceXSmall = 5.0;
-  static const double _HorizontalSpaceSmall = 10.0;
-  static const double _HorizontalSpaceMedium = 20.0;
-  static const double _HorizontalSpaceLarge = 60.0;
-
-  static const Widget verticalSpaceXSmall =
-      SizedBox(height: _VerticalSpaceXSmall);
-  static const Widget verticalSpaceSmall =
-      SizedBox(height: _VerticalSpaceSmall);
-  static const Widget verticalSpaceMedium =
-      SizedBox(height: _VerticalSpaceMedium);
-  static const Widget verticalSpaceLarge =
-      SizedBox(height: _VerticalSpaceLarge);
-
-  static const Widget horizontalSpaceXSmall =
-      SizedBox(width: _HorizontalSpaceXSmall);
-  static const Widget horizontalSpaceSmall =
-      SizedBox(width: _HorizontalSpaceSmall);
-  static const Widget horizontalSpaceMedium =
-      SizedBox(width: _HorizontalSpaceMedium);
-  static const Widget horizontalSpaceLarge =
-      SizedBox(width: _HorizontalSpaceLarge);
-}
