+++ lib/view/screens/blank_screen.dart
@@ -1,14 +1,6 @@
-import 'package:firebase_shop_app/core/providers/cart_provider.dart';
-import 'package:firebase_shop_app/core/providers/product_model_provider.dart';
-import 'package:firebase_shop_app/core/providers/products_provider.dart';
-import 'package:firebase_shop_app/utils/view/constant_routs.dart';
-import 'package:firebase_shop_app/view/shared/badge.dart';
-import 'package:firebase_shop_app/view/widgets/product_details_widget.dart';
 import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
 
 class BlankScreen extends StatelessWidget {
-
   @override
   Widget build(BuildContext context) {
     return Container();
