--- lib/view/screens/orders_screen.dart
@@ -1,20 +1,0 @@
-import 'package:firebase_shop_app/core/providers/orders_provider.dart';
-import 'package:firebase_shop_app/view/widgets/order_item_widget.dart';
-import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
-
-class OrdersScreen extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    final orderData = Provider.of<OrdersProvider>(context);
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('Moje Zamówienia'),
-      ),
-      body: ListView.builder(
-        itemCount: orderData.orders.length,
-        itemBuilder: (ctx, i) => OrderItemWidget(orderData.orders[i]),
-      ),
-    );
-  }
-}
