--- lib/view/screens/un_defined_screen.dart
@@ -1,16 +1,0 @@
-import 'package:firebase_shop_app/view/widgets/bottom_app_bar.dart';
-import 'package:flutter/material.dart';
-
-class UndefinedScreen extends StatelessWidget {
-  final String name;
-  const UndefinedScreen({Key key, this.name}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: Center(
-        child: Text('Route for $name is not defined'),
-      ),
-    );
-  }
-}
