+++ lib/view/shared/custom_bottom_app_bar.dart
@@ -1,0 +1,101 @@
+import 'package:flutter/material.dart';
+
+class BottomAppBarItem {
+  BottomAppBarItem({this.iconData, this.text});
+
+  IconData iconData;
+  String text;
+}
+
+class CustomBottomAppBar extends StatefulWidget {
+  CustomBottomAppBar({
+    this.items,
+    this.centerItemText,
+    this.height: 60.0,
+    this.iconSize: 24.0,
+    this.backgroundColor,
+    this.color,
+    this.notchedShape,
+    this.onTabSelected,
+    Color selectedColor,
+  }) {
+    Color(0xFFFA6442);
+  }
+
+  final List<BottomAppBarItem> items;
+  final String centerItemText;
+  final double height;
+  final double iconSize;
+  final Color backgroundColor;
+  final Color color;
+  final Color selectedColor = Color(0xFFFA6442);
+  final NotchedShape notchedShape;
+  final ValueChanged<int> onTabSelected;
+
+  @override
+  State<StatefulWidget> createState() => CustomBottomAppBarState();
+}
+
+class CustomBottomAppBarState extends State<CustomBottomAppBar> {
+  int _selectedIndex = 3;
+
+  _updateIndex(int index) {
+    widget.onTabSelected(index);
+    setState(() {
+      _selectedIndex = index;
+    });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    List<Widget> items = List.generate(widget.items.length, (int index) {
+      return _buildTabItem(
+        item: widget.items[index],
+        index: index,
+        onPressed: _updateIndex,
+      );
+    });
+    // items.insert(items.length >> 1, _buildMiddleTabItem());
+
+    return BottomAppBar(
+      shape: widget.notchedShape,
+      child: Row(
+        textDirection: TextDirection.rtl,
+        mainAxisSize: MainAxisSize.max,
+        mainAxisAlignment: MainAxisAlignment.spaceAround,
+        children: items,
+      ),
+      color: widget.backgroundColor,
+    );
+  }
+
+  Widget _buildTabItem({
+    BottomAppBarItem item,
+    int index,
+    ValueChanged<int> onPressed,
+  }) {
+    Color color = _selectedIndex == index ? widget.selectedColor : widget.color;
+    return Expanded(
+      child: SizedBox(
+        height: widget.height,
+        child: Material(
+          type: MaterialType.transparency,
+          child: InkWell(
+            onTap: () => onPressed(index),
+            child: Column(
+              mainAxisSize: MainAxisSize.min,
+              mainAxisAlignment: MainAxisAlignment.center,
+              children: <Widget>[
+                Icon(item.iconData, color: color, size: widget.iconSize),
+                Text(
+                  item.text,
+                  style: TextStyle(color: color, fontSize: 12),
+                )
+              ],
+            ),
+          ),
+        ),
+      ),
+    );
+  }
+}
