+++ lib/view/widgets/bottom_app_bar.dart
@@ -1,13 +1,7 @@
-import 'package:firebase_shop_app/utils/view/constant_routs.dart';
+import 'package:firebase_shop_app/view/screens/all_products_screen.dart';
 import 'package:firebase_shop_app/view/screens/blank_screen.dart';
-import 'package:firebase_shop_app/view/screens/cart_screen.dart';
-import 'package:firebase_shop_app/view/screens/favourite_screen.dart';
-import 'package:firebase_shop_app/view/screens/more_screen.dart';
 import 'package:firebase_shop_app/view/screens/my_products_screen.dart';
-import 'package:firebase_shop_app/view/screens/orders_screen.dart';
-import 'package:firebase_shop_app/view/screens/products_overview_screen.dart';
-import 'package:firebase_shop_app/view/screens/un_defined_screen.dart';
-import 'package:firebase_shop_app/view/shared/custom_bottomAppBar.dart';
+import 'package:firebase_shop_app/view/shared/custom_bottom_app_bar.dart';
 import 'package:flutter/material.dart';
 
 class SharedBottomAppBar extends StatefulWidget {
@@ -16,24 +10,18 @@ class SharedBottomAppBar extends StatefulWidget {
 }
 
 class _SharedBottomAppBarState extends State<SharedBottomAppBar> {
-  Widget _lastSelected = ProductsOverviewScreen();
+  Widget _lastSelected = AllProductsScreen();
 
-  String _title = 'Strona główna';
   List<Widget> pages = [
-
     BlankScreen(),
     MyProductsScreen(),
-    //MoreScreen(),
-    //OrdersScreen(),
-    ProductsOverviewScreen(),
+    AllProductsScreen(),
   ];
   List<String> titles = ['Wiadomości', 'Wystaw produkt', 'Strona główna'];
 
   void _selectedTab(int index) {
     setState(() {
-      print(index);
       _lastSelected = pages[index];
-      _title = titles[index];
     });
   }
