+++ pubspec.yaml
@@ -32,9 +32,8 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  firebase_core: ^0.5.2
-  firebase_database: ^4.3.0
-
+  firebase_core: ^0.4.1+3
+  cloud_firestore: ^0.12.10
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
